+++ MatrixSDK/Crypto/CryptoMachine/MXCryptoMachine.swift
@@ -58,6 +58,7 @@ class MXCryptoMachine {
     
     private let machine: OlmMachine
     private let requests: MXCryptoRequests
+    private let queryScheduler: MXKeysQueryScheduler<MXKeysQueryResponse>
     private let getRoomAction: GetRoomAction
     
     private let sessionsQueue = MXTaskQueue()
@@ -72,28 +73,39 @@ class MXCryptoMachine {
     
     private let log = MXNamedLog(name: "MXCryptoMachine")
 
-    init(userId: String, deviceId: String, restClient: MXRestClient, getRoomAction: @escaping GetRoomAction) throws {
+    init(
+        userId: String,
+        deviceId: String,
+        restClient: MXRestClient,
+        getRoomAction: @escaping GetRoomAction
+    ) throws {
         let url = try Self.storeURL(for: userId)
+        
         machine = try OlmMachine(
             userId: userId,
             deviceId: deviceId,
             path: url.path,
             passphrase: nil
         )
-        requests = MXCryptoRequests(restClient: restClient)
+        let requests = MXCryptoRequests(restClient: restClient)
+        self.requests = requests
+        
+        queryScheduler = MXKeysQueryScheduler { users in
+            try await requests.queryKeys(users: users)
+        }
         self.getRoomAction = getRoomAction
         
-        setLogger(logger: self)
-    }
-    
-    func start() async throws {
         let details = """
-        Starting the crypto machine for \(userId)
+        Initialized the crypto machine for \(userId)
           - device id  : \(deviceId)
           - ed25519    : \(deviceEd25519Key ?? "")
           - curve25519 : \(deviceCurve25519Key ?? "")
         """
         log.debug(details)
+    }
+    
+    func uploadKeysIfNecessary() async throws {
+        log.debug("Checking for keys to upload")
         
         var keysUploadRequest: Request?
         for request in try machine.outgoingRequests() {
@@ -110,7 +122,6 @@ class MXCryptoMachine {
         }
         
         try await handleRequest(request)
-        
         log.debug("Keys successfully uploaded")
     }
     
@@ -178,6 +189,8 @@ extension MXCryptoMachine: MXCryptoIdentity {
 }
 
 extension MXCryptoMachine: MXCryptoSyncing {
+    
+    @MainActor
     func handleSyncResponse(
         toDevice: MXToDeviceSyncResponse?,
         deviceLists: MXDeviceListResponse?,
@@ -211,6 +224,31 @@ extension MXCryptoMachine: MXCryptoSyncing {
         return toDevice
     }
     
+    func downloadKeysIfNecessary(users: [String]) async throws {
+        machine.updateTrackedUsers(users: users)
+        try await withThrowingTaskGroup(of: Void.self) { [weak self] group in
+            guard let self = self else { return }
+
+            for request in try machine.outgoingRequests() {
+                if case .keysQuery(_, let requestUsers) = request {
+                    let usersInCommon = Set(requestUsers).intersection(users)
+                    if !usersInCommon.isEmpty {
+                        try await self.handleRequest(request)
+                        return
+                    }
+                }
+            }
+        }
+    }
+    
+    @available(*, deprecated, message: "The application should not manually force reload keys, use `downloadKeysIfNecessary` instead")
+    func reloadKeys(users: [String]) async throws {
+        machine.updateTrackedUsers(users: users)
+        try await handleRequest(
+            .keysQuery(requestId: UUID().uuidString, users: users)
+        )
+    }
+    
     func processOutgoingRequests() async throws {
         try await syncQueue.sync { [weak self] in
             try await self?.handleOutgoingRequests()
@@ -234,7 +272,8 @@ extension MXCryptoMachine: MXCryptoSyncing {
             try markRequestAsSent(requestId: requestId, requestType: .keysUpload, response: response.jsonString())
             
         case .keysQuery(let requestId, let users):
-            let response = try await requests.queryKeys(users: users)
+            // Key queries go through a scheduler layer instead of directly through the rest client
+            let response = try await queryScheduler.query(users: Set(users))
             try markRequestAsSent(requestId: requestId, requestType: .keysQuery, response: response.jsonString())
             
         case .keysClaim(let requestId, let oneTimeKeys):
@@ -341,21 +380,6 @@ extension MXCryptoMachine: MXCryptoUserIdentitySource {
         }
     }
     
-    func isUserTracked(userId: String) -> Bool {
-        do {
-            return try machine.isUserTracked(userId: userId)
-        } catch {
-            log.error("Failed checking user tracking")
-            return false
-        }
-    }
-    
-    func downloadKeys(users: [String]) async throws {
-        try await handleRequest(
-            .keysQuery(requestId: UUID().uuidString, users: users)
-        )
-    }
-    
     func verifyUser(userId: String) async throws {
         let request = try machine.verifyIdentity(userId: userId)
         try await requests.uploadSignatures(request: request)
@@ -372,13 +396,17 @@ extension MXCryptoMachine: MXCryptoUserIdentitySource {
 }
 
 extension MXCryptoMachine: MXCryptoRoomEventEncrypting {
+    func addTrackedUsers(_ users: [String]) {
+        machine.updateTrackedUsers(users: users)
+    }
+    
     func shareRoomKeysIfNecessary(
         roomId: String,
         users: [String],
         settings: EncryptionSettings
     ) async throws {
         try await sessionsQueue.sync { [weak self] in
-            try await self?.updateTrackedUsers(users: users)
+            try await self?.downloadKeysIfNecessary(users: users)
             try await self?.getMissingSessions(users: users)
         }
         
@@ -410,25 +438,6 @@ extension MXCryptoMachine: MXCryptoRoomEventEncrypting {
     }
     
     // MARK: - Private
-    
-    private func updateTrackedUsers(users: [String]) async throws {
-        machine.updateTrackedUsers(users: users)
-        try await withThrowingTaskGroup(of: Void.self) { [weak self] group in
-            guard let self = self else { return }
-
-            for request in try machine.outgoingRequests() {
-                guard case .keysQuery = request else {
-                    continue
-                }
-
-                group.addTask {
-                    try await self.handleRequest(request)
-                }
-            }
-
-            try await group.waitForAll()
-        }
-    }
     
     private func getMissingSessions(users: [String]) async throws {
         guard
@@ -466,7 +475,7 @@ extension MXCryptoMachine: MXCryptoRoomEventDecrypting {
             log.failure("Invalid event")
             throw Error.invalidEvent
         }
-        return try machine.decryptRoomEvent(event: eventString, roomId: roomId)
+        return try machine.decryptRoomEvent(event: eventString, roomId: roomId, handleVerificatonEvents: true)
     }
     
     func requestRoomKey(event: MXEvent) async throws {
@@ -484,6 +493,10 @@ extension MXCryptoMachine: MXCryptoRoomEventDecrypting {
 }
 
 extension MXCryptoMachine: MXCryptoCrossSigning {
+    func refreshCrossSigningStatus() async throws {
+        try await reloadKeys(users: [userId])
+    }
+    
     func crossSigningStatus() -> CrossSigningStatus {
         return machine.crossSigningStatus()
     }
@@ -741,12 +754,6 @@ extension MXCryptoMachine: MXCryptoBackup {
             cachedRoomKeyCounts = nil
         }
         isComputingRoomKeyCounts = false
-    }
-}
-
-extension MXCryptoMachine: Logger {
-    func log(logLine: String) {
-        MXLog.debug("[MXCryptoMachine] \(logLine)")
     }
 }