+++ MatrixSDK/MXSDKOptions.m
@@ -56,12 +56,13 @@ static MXSDKOptions *sharedOnceInstance = nil;
         _enableRoomSharedHistoryOnInvite = NO;
         
         #if DEBUG
-        _enableCryptoV2 = NO;
+        _isCryptoSDKAvailable = NO;
+        _enableCryptoSDK = NO;
         #endif
         
         _enableSymmetricBackup = NO;
         _enableNewClientInformationFeature = NO;
-        _enableSyncProgress = NO;
+        _enableStartupProgress = NO;
     }
     
     return self;